# copilot/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookYourCoach API",
        "description": "API REST pour la plateforme de réservation de cours avec coaches équestres ou autres sports. Cette API permet la gestion complète des utilisateurs (admin/enseignants/élèves), des réservations, des paiements et de la facturation.",
        "contact": {
            "name": "BookYourCoach Support",
            "url": "https://bookyourcoach.com",
            "email": "support@bookyourcoach.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur de développement BookYourCoach"
        }
    ],
    "paths": {
...
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Liste des profils",
                "description": "Retourne la liste de tous les profils utilisateurs",
                "operationId": "getProfilesList",
                "responses": {
                    "200": {
                        "description": "Liste des profils récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "profiles": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Profile"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Créer un profil utilisateur",
                "description": "Crée un nouveau profil pour l'utilisateur authentifié",
                "operationId": "createProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+33123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue de la Paix"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "75001"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    },
                                    "emergency_contact_name": {
                                        "type": "string",
                                        "example": "Marie Dupont"
                                    },
                                    "emergency_contact_phone": {
                                        "type": "string",
                                        "example": "+33987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile created successfully"
                                        },
                                        "profile": {
                                            "$ref": "#/components/schemas/Profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profiles/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Détails d'un profil",
                "description": "Retourne les détails d'un profil spécifique",
                "operationId": "getProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du profil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "profile": {
                                            "$ref": "#/components/schemas/Profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Modifier un profil",
                "description": "Met à jour les informations d'un profil",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "profile": {
                                            "$ref": "#/components/schemas/Profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Supprimer un profil",
                "description": "Supprime un profil utilisateur",
                "operationId": "deleteProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "title": "Profile",
                "description": "Profil utilisateur avec informations personnelles et de contact",
                "properties": {
                    "id": {
                        "description": "Identifiant unique du profil",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur associé",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "description": "Prénom",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Jean"
                    },
                    "last_name": {
                        "description": "Nom de famille",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Dupont"
                    },
                    "phone": {
                        "description": "Numéro de téléphone",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456789"
                    },
                    "address": {
                        "description": "Adresse complète",
                        "type": "string",
                        "example": "123 Rue de la Paix"
                    },
                    "city": {
                        "description": "Ville",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Paris"
                    },
                    "postal_code": {
                        "description": "Code postal",
                        "type": "string",
                        "maxLength": 10,
                        "example": "75001"
                    },
                    "country": {
                        "description": "Pays",
                        "type": "string",
                        "maxLength": 100,
                        "example": "France"
                    },
                    "date_of_birth": {
                        "description": "Date de naissance",
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15"
                    },
                    "emergency_contact_name": {
                        "description": "Nom du contact d'urgence",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Marie Dupont"
                    },
                    "emergency_contact_phone": {
                        "description": "Téléphone du contact d'urgence",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33987654321"
                    },
                    "medical_notes": {
                        "description": "Notes médicales",
                        "type": "string",
                        "example": "Allergie aux abeilles"
                    },
                    "preferences": {
                        "description": "Préférences utilisateur (JSON)",
                        "type": "object",
                        "example": {
                            "language": "fr",
                            "notifications": true
                        }
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Bearer. Utilisez le token obtenu via /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}